modifications to run hsd
