+++ nupic/research/frameworks/dynamic_sparse/networks/__init__.py
@@ -3,3 +3,4 @@ from .gsc import *
 from .hebbian import *
 from .wideresnet import *
 from .layers import *
+from .utils import *
