+++ projects/dynamic_sparse/runs/replicate_hsd.py
@@ -44,8 +44,8 @@ base_exp_config = dict(
     network="GSCHeb",
     percent_on_k_winner=[0.095, 0.125, 0.067],
     k_inference_factor=1.5,
-    boost_strength = 1.5,
-    boost_strength_factor = 0.9,
+    boost_strength=[1.5, 1.5, 1.5],
+    boost_strength_factor=[0.9, 0.9, 0.9],
     hidden_neurons_conv=[64, 64],
     hidden_neurons_fc=1500,    
     bias=True,
@@ -79,7 +79,7 @@ tune_config = dict(
     checkpoint_freq=0,
     checkpoint_at_end=False,
     stop={"training_iteration": 25},
-    resources_per_trial={"cpu": 1, "gpu": .25},
+    resources_per_trial={"cpu": 1, "gpu": 1},
     loggers=DEFAULT_LOGGERS,
     verbose=0,
 )
